Adding updated script to convert model from .h5 to .pb from CMSSW use. Also adding trainings 0329 and 0407
